Create eighth_program.py
Change- Program that asks how much purchases cost and the amount of paid money and then
prints the amount of change. To simplify the program, only 1, 2, 5 and 10 euros
are offered as change. It is assumed that the paid amount is always euros, i.e.
no cents are paid besides euros. It is further assumed that the paid amount is
always at least 1 euro.