Create program 3_8_2.py
The function read_input, which reads the number entered by the user, checks that it is larger than zero, and returns it to the caller of the function. The function also requests the user to enter a new input until the user enters a positive number. Add the print_box function printing the box.